// src/api/ix/client/types.gen.ts
export type PostV2InvoicesData = {
    body?: {
        invoice: {
            date: string | string;
            due_date: string | string;
            tax_exemption_reason?: string;
            client: {
                id?: number;
                name?: string;
                email?: string;
                fiscal_id?: string;
                address?: string;
                postal_code?: string;
                country?: string;
                code?: string | null;
                city?: string;
                phone?: string;
            };
            observations?: string;
            items: Array<{
                quantity: number;
                id?: number;
                name?: string;
                code?: string;
                description?: string;
                tax: {
                    id?: number;
                    name: string;
                    value: number;
                } | number;
                unit_price: number;
                rrp?: number;
                unit?: string | null;
                retention?: number;
                discount?: number;
                discount_amount?: number;
                subtotal?: number;
                tax_amount?: number;
                total?: number;
            }>;
            global_discount?: {
                value_type: 'percentage' | 'absolute';
                value: number;
            };
            reference?: string;
            retention?: string;
            manual_sequence_number?: string;
        };
    };
    headers: {
        'x-account-name': string;
        'x-api-key': string;
        'x-env'?: 'prod' | 'dev';
    };
    path?: never;
    query?: {
        /**
         * Comma-separated resolvers. Supported values: on_tax_fallback_search_tax_by_value
         */
        resolvers?: string;
    };
    url: '/v2/invoices';
};
// ...
export type PostV2CreditNotesData = {
    body?: {
        credit_note: {
            date: string | string;
            due_date: string | string;
            tax_exemption_reason?: string;
            client: {
                id?: number;
                name?: string;
                email?: string;
                fiscal_id?: string;
                address?: string;
                postal_code?: string;
                country?: string;
                code?: string | null;
                city?: string;
                phone?: string;
            };
            observations?: string;
            items: Array<{
                quantity: number;
                id?: number;
                name?: string;
                code?: string;
                description?: string;
                tax: {
                    id?: number;
                    name: string;
                    value: number;
                } | number;
                unit_price: number;
                rrp?: number;
                unit?: string | null;
                retention?: number;
                discount?: number;
                discount_amount?: number;
                subtotal?: number;
                tax_amount?: number;
                total?: number;
            }>;
            global_discount?: {
                value_type: 'percentage' | 'absolute';
                value: number;
            };
            reference?: string;
            retention?: string;
            manual_sequence_number?: string;
            owner_invoice_id: number;
        };
    };
    headers: {
        'x-account-name': string;
        'x-api-key': string;
        'x-env'?: 'prod' | 'dev';
    };
    path?: never;
    query?: {
        /**
         * Comma-separated resolvers. Supported values: on_tax_fallback_search_tax_by_value
         */
        resolvers?: string;
    };
    url: '/v2/credit_notes';
};